001--scratch/matlab/gen_unique_two_terminal_core_graphs.m - using an invariant sorted degree sequence as classification invariant to reduce the search space for isomorphic graphs. * Function to generate connected simple graphs with no dangling branches. * No dangling branches means that every internal (non-terminal) node lies on at least one path through the terminals.
Number of unique two-terminal core graphs with 2 to 6 nodes and 1 to 5 edges: Elapsed time is 2.390838 seconds.
Number of unique two-terminal core graphs with 2 to 7 nodes and 1 to 6 edges: Elapsed time is 35.710352 seconds.
Number of unique two-terminal core graphs with 2 to 8 nodes and 1 to 7 edges: Elapsed time is 843.135090 seconds.
Number of unique two-terminal core graphs with 2 to 9 nodes and 1 to 8 edges: Elapsed time is 18693.755631 seconds.
---
Number of unique two-terminal core graphs (optimized) with 2 to 6 nodes and 1 to 5 edges: Elapsed time is 2.337227 seconds.
Number of unique two-terminal core graphs (optimized) with 2 to 7 nodes and 1 to 6 edges: Elapsed time is 34.255558 seconds.
Number of unique two-terminal core graphs (optimized) with 2 to 8 nodes and 1 to 7 edges: Elapsed time is 747.454588 seconds.

See #1 (Scratch the surface with some MATLAB scripts)